backend/requirements.txt
matplotlib>=3.8,<4.0
scikit-learn==1.3.2
joblib==1.3.2
fastapi==0.108.0
uvicorn==0.25.0
# pydantic-settings==2.1.0
pydantic-settings>=0.1
# pydantic-settings>=0.1
# psycopg2-binary==2.9.3
psycopg2-binary>=2.9.3
# pydantic>=1.8,<2
python-jose==3.3.0
apitally[fastapi]==0.3.3
starlette==0.32.0
alembic==1.13.1
SQLAlchemy==1.4.0
# psycopg2 ==2.8.6
##################
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
botocore==1.34.69
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
fsspec==2024.3.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
numpy==1.26.4
pandas >= 2.2, <3.0
pyarrow==15.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3fs==2024.3.1
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4